Density Profiles: Rename Dir & Namespaces
