+++ src/picongpu/include/particles/densityProfiles/FreeFormulaImpl.def
@@ -18,16 +18,14 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-
 #pragma once
+
 
 namespace picongpu
 {
-
-namespace gasProfiles
+namespace densityProfiles
 {
-
-struct HomogenousImpl;
-
-} //namespace gasProfiles
-} //namespace picongpu
+    template<typename T_ParamClass>
+    struct FreeFormulaImpl;
+}
+}
