+++ src/picongpu/include/particles/densityProfiles/FromHDF5Impl.def
@@ -1,5 +1,5 @@
 /**
- * Copyright 2013-2017 Axel Huebl, Heiko Burau, Rene Widera, Felix Schmitt
+ * Copyright 2014-2017 Rene Widera
  *
  * This file is part of PIConGPU.
  *
@@ -18,33 +18,14 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-
 #pragma once
 
-#include "simulation_defines.hpp"
-#include "particles/gasProfiles/IProfile.def"
 
 namespace picongpu
 {
-
-namespace gasProfiles
+namespace densityProfiles
 {
-
-template<typename T_Base>
-struct IProfile : private T_Base
-{
-
-    typedef T_Base Base;
-
-    HINLINE IProfile(uint32_t currentStep) : Base(currentStep)
-    {
-    }
-
-    HDINLINE float_X operator()(const DataSpace<simDim>& totalCellOffset)
-    {
-        return Base::operator()(totalCellOffset);
-    }
-};
-
-} //namespace gasProfiles
-} //namespace picongpu
+    template<typename T_ParamClass>
+    struct FromHDF5Impl;
+}
+}
