+++ src/picongpu/include/particles/densityProfiles/GaussianCloudImpl.def
@@ -18,17 +18,14 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-
 #pragma once
+
 
 namespace picongpu
 {
-
-namespace gasProfiles
+namespace densityProfiles
 {
-
-template<typename T_ParamClass>
-struct FromHDF5Impl;
-
-} //namespace gasProfiles
-} //namespace picongpu
+    template<typename T_ParamClass>
+    struct GaussianCloudImpl;
+}
+}
