+++ src/picongpu/include/particles/densityProfiles/HomogenousImpl.def
@@ -18,17 +18,13 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-
 #pragma once
+
 
 namespace picongpu
 {
-
-namespace gasProfiles
+namespace densityProfiles
 {
-
-template<typename T_ParamClass>
-struct SphereFlanksImpl;
-
-} //namespace gasProfiles
-} //namespace picongpu
+    struct HomogenousImpl;
+}
+}
