+++ src/picongpu/include/particles/densityProfiles/LinearExponentialImpl.def
@@ -18,17 +18,14 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-
 #pragma once
+
 
 namespace picongpu
 {
-
-namespace gasProfiles
+namespace densityProfiles
 {
-
-template<typename T_ParamClass>
-struct LinearExponentialImpl;
-
-} //namespace gasProfiles
-} //namespace picongpu
+    template<typename T_ParamClass>
+    struct LinearExponentialImpl;
+}
+}
