+++ src/picongpu/include/particles/densityProfiles/profiles.def
@@ -1,5 +1,5 @@
 /**
- * Copyright 2014-2017 Rene Widera
+ * Copyright 2014-2017 Rene Widera, Alexander Grund
  *
  * This file is part of PIConGPU.
  *
@@ -18,17 +18,13 @@
  * If not, see <http://www.gnu.org/licenses/>.
  */
 
-
 #pragma once
 
-namespace picongpu
-{
-
-namespace gasProfiles
-{
-
-template<typename T_ParamClass>
-struct GaussianCloudImpl;
-
-} //namespace gasProfiles
-} //namespace picongpu
+#include "particles/densityProfiles/IProfile.def"
+#include "particles/densityProfiles/FreeFormulaImpl.def"
+#include "particles/densityProfiles/GaussianImpl.def"
+#include "particles/densityProfiles/HomogenousImpl.def"
+#include "particles/densityProfiles/LinearExponentialImpl.def"
+#include "particles/densityProfiles/GaussianCloudImpl.def"
+#include "particles/densityProfiles/SphereFlanksImpl.def"
+#include "particles/densityProfiles/FromHDF5Impl.def"
